revising code in gradle plugin
